+++ p2p/providers/addrbook_provider/addrbook_provider.go
@@ -17,8 +17,7 @@ const ModuleName = "addrbook_provider"
 
 // AddrBookProvider is an interface that provides AddrBook accessors
 type AddrBookProvider interface {
-	modules.IntegratableModule
-	modules.InterruptableModule
+	modules.Module
 
 	GetStakedAddrBookAtHeight(height uint64) (typesP2P.AddrBook, error)
 	GetConnFactory() typesP2P.ConnectionFactory