+++ p2p/types/network_peer.go
@@ -1,6 +1,8 @@
 package types
 
 import (
+	"github.com/multiformats/go-multiaddr"
+
 	cryptoPocket "github.com/pokt-network/pocket/shared/crypto"
 )
 
@@ -8,6 +10,7 @@ type NetworkPeer struct {
 	Dialer    Transport
 	PublicKey cryptoPocket.PublicKey
 	Address   cryptoPocket.Address
+	Multiaddr multiaddr.Multiaddr
 
 	// This is only included because it's a more human-friendly differentiator between peers
 	ServiceURL string