+++ app/client/doc/CHANGELOG.md
@@ -7,10 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.0.0.20] - 2023-02-28
+## [0.0.0.21] - 2023-03-08
 
 - Introduces vault keybase to allow for the use of a vault server to store the private keys
 - Fixes bug where private keys were not being properly parsed from the debug keybase
+
+## [0.0.0.20] - 2023-03-03
+
+- Support libp2p module in debug CLI
 
 ## [0.0.0.19] - 2023-02-28