+++ p2p/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.33] - 2023-03-03
+
+- Add TECHDEBT comments
+
+## [0.0.0.32] - 2023-03-03
+
+- Added embedded `modules.InitializableModule` to the P2P `AddrBookProvider` interface so that it can be dependency injected as a `modules.Module` via the bus registry.
+
 ## [0.0.0.31] - 2023-03-01
 
 - replace `consensus_port` with `port` in P2P config